Replace relative links with absolute
Some of the links in README.md were relative to the repo. That works fine for the repo page in GitHub, but they result in 404 errors on the generated github.io website. There isn't a great solution, but absolute links will work as long as the repo site isn't changed.